// x-pack/platform/plugins/shared/cases/common/workflows/steps/create_case.d.ts
export declare const InputSchema: z.ZodObject<{
    tags: z.ZodOptional<z.ZodArray<z.ZodString>>;
    description: z.ZodString;
    settings: z.ZodOptional<z.ZodObject<{
        syncAlerts: z.ZodBoolean;
        extractObservables: z.ZodOptional<z.ZodBoolean>;
    }, z.core.$strip>>;
    title: z.ZodString;
    severity: z.ZodOptional<z.ZodEnum<{
        critical: "critical";
        medium: "medium";
        high: "high";
        low: "low";
    }>>;
    category: z.ZodOptional<z.ZodString>;
    owner: z.ZodEnum<{
        cases: "cases";
        observability: "observability";
        securitySolution: "securitySolution";
    }>;
    assignees: z.ZodOptional<z.ZodNullable<z.ZodArray<z.ZodObject<{
        uid: z.ZodString;
    }, z.core.$strip>>>>;
    customFields: z.ZodOptional<z.ZodArray<z.ZodObject<{
        key: z.ZodString;
        type: z.ZodEnum<{
            text: "text";
            toggle: "toggle";
        }>;
        value: z.ZodUnion<readonly [z.ZodNullable<z.ZodString>, z.ZodBoolean]>;
    }, z.core.$strip>>>;
}, z.core.$strip>;
export declare const OutputSchema: z.ZodObject<{
    case: z.ZodObject<{
// ...
        comments: z.ZodArray<z.ZodDiscriminatedUnion<[z.ZodObject<{
// ...
            owner: z.ZodOptional<z.ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, z.core.$strip>;
}, z.core.$strip>;
declare const ConfigSchema: z.ZodObject<{
    'push-case': z.ZodOptional<z.ZodDefault<z.ZodOptional<z.ZodBoolean>>>;
    'connector-id': z.ZodOptional<z.ZodString>;
}, z.core.$strip>;
type CreateCaseStepInputSchema = typeof InputSchema;
type CreateCaseStepOutputSchema = typeof OutputSchema;
export type CreateCaseStepInput = z.infer<typeof InputSchema>;
export type CreateCaseStepOutput = z.infer<typeof OutputSchema>;
export type CreateCaseStepConfig = z.infer<typeof ConfigSchema>;
export declare const createCaseStepCommonDefinition: CommonStepDefinition<CreateCaseStepInputSchema, CreateCaseStepOutputSchema>;