// x-pack/platform/plugins/shared/cases/common/workflows/steps/create_case_from_template.d.ts
declare const InputSchema: z.ZodObject<{
    owner: z.ZodEnum<{
        cases: "cases";
        observability: "observability";
        securitySolution: "securitySolution";
    }>;
    case_template_id: z.ZodString;
    overwrites: z.ZodOptional<z.ZodObject<{
        tags: z.ZodOptional<z.ZodArray<z.ZodString>>;
        status: z.ZodOptional<z.ZodEnum<{
            closed: "closed";
            open: "open";
            "in-progress": "in-progress";
        }>>;
        description: z.ZodOptional<z.ZodString>;
        connector: z.ZodOptional<z.ZodUnion<readonly [z.ZodObject<{
            fields: z.ZodNullable<z.ZodString>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".none">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodNullable<z.ZodString>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".cases-webhook">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodObject<{
                issueType: z.ZodNullable<z.ZodString>;
                parent: z.ZodNullable<z.ZodString>;
                priority: z.ZodNullable<z.ZodString>;
            }, z.core.$strip>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".jira">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodNullable<z.ZodObject<{
                issueTypes: z.ZodArray<z.ZodString>;
                severityCode: z.ZodString;
            }, z.core.$strip>>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".resilient">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodObject<{
                category: z.ZodNullable<z.ZodString>;
                impact: z.ZodNullable<z.ZodString>;
                severity: z.ZodNullable<z.ZodString>;
                subcategory: z.ZodNullable<z.ZodString>;
                urgency: z.ZodNullable<z.ZodString>;
            }, z.core.$strip>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".servicenow">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodObject<{
                category: z.ZodNullable<z.ZodString>;
                destIp: z.ZodNullable<z.ZodBoolean>;
                malwareHash: z.ZodNullable<z.ZodBoolean>;
                malwareUrl: z.ZodNullable<z.ZodBoolean>;
                priority: z.ZodNullable<z.ZodString>;
                sourceIp: z.ZodNullable<z.ZodBoolean>;
                subcategory: z.ZodNullable<z.ZodString>;
            }, z.core.$strip>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".servicenow-sir">;
        }, z.core.$strip>, z.ZodObject<{
            fields: z.ZodObject<{
                caseId: z.ZodNullable<z.ZodString>;
            }, z.core.$strip>;
            id: z.ZodString;
            name: z.ZodString;
            type: z.ZodLiteral<".swimlane">;
        }, z.core.$strip>]>>;
        settings: z.ZodOptional<z.ZodObject<{
            syncAlerts: z.ZodBoolean;
            extractObservables: z.ZodOptional<z.ZodBoolean>;
        }, z.core.$strip>>;
        title: z.ZodOptional<z.ZodString>;
        severity: z.ZodOptional<z.ZodEnum<{
            critical: "critical";
            medium: "medium";
            high: "high";
            low: "low";
        }>>;
        category: z.ZodOptional<z.ZodString>;
        assignees: z.ZodOptional<z.ZodNullable<z.ZodArray<z.ZodObject<{
            uid: z.ZodString;
        }, z.core.$strip>>>>;
        customFields: z.ZodOptional<z.ZodArray<z.ZodObject<{
            key: z.ZodString;
            type: z.ZodEnum<{
                text: "text";
                toggle: "toggle";
            }>;
            value: z.ZodUnion<readonly [z.ZodNullable<z.ZodString>, z.ZodBoolean]>;
        }, z.core.$strip>>>;
        closeReason: z.ZodOptional<z.ZodUnion<readonly [z.ZodEnum<{
            other: "other";
            duplicate: "duplicate";
            false_positive: "false_positive";
            true_positive: "true_positive";
            benign_positive: "benign_positive";
            automated_closure: "automated_closure";
        }>, z.ZodString]>>;
    }, z.core.$strip>>;
}, z.core.$strip>;
declare const OutputSchema: z.ZodObject<{
    case: z.ZodObject<{
// ...
        comments: z.ZodArray<z.ZodDiscriminatedUnion<[z.ZodObject<{
// ...
            owner: z.ZodOptional<z.ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, z.core.$strip>;
}, z.core.$strip>;
declare const ConfigSchema: z.ZodObject<{
    'push-case': z.ZodOptional<z.ZodDefault<z.ZodOptional<z.ZodBoolean>>>;
}, z.core.$strip>;
type CreateCaseFromTemplateStepInputSchema = typeof InputSchema;
type CreateCaseFromTemplateStepOutputSchema = typeof OutputSchema;
export type CreateCaseFromTemplateStepInput = z.infer<typeof InputSchema>;
export type CreateCaseFromTemplateStepOutput = z.infer<typeof OutputSchema>;
export type CreateCaseFromTemplateStepConfig = z.infer<typeof ConfigSchema>;
export declare const createCaseFromTemplateStepCommonDefinition: CommonStepDefinition<CreateCaseFromTemplateStepInputSchema, CreateCaseFromTemplateStepOutputSchema>;