// x-pack/platform/plugins/shared/cases/common/workflows/steps/assign_case.d.ts
declare const InputSchema: z.ZodObject<{
    case_id: z.ZodString;
    version: z.ZodOptional<z.ZodString>;
    assignees: z.ZodNullable<z.ZodArray<z.ZodObject<{
        uid: z.ZodString;
    }, z.core.$strip>>>;
}, z.core.$strip>;
declare const OutputSchema: z.ZodObject<{
    case: z.ZodObject<{
// ...
        comments: z.ZodArray<z.ZodDiscriminatedUnion<[z.ZodObject<{
// ...
            owner: z.ZodOptional<z.ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, z.core.$strip>;
}, z.core.$strip>;
type AssignCaseStepInputSchema = typeof InputSchema;
type AssignCaseStepOutputSchema = typeof OutputSchema;
export type AssignCaseStepInput = z.infer<typeof InputSchema>;
export declare const assignCaseStepCommonDefinition: CommonStepDefinition<AssignCaseStepInputSchema, AssignCaseStepOutputSchema>;