// x-pack/platform/plugins/shared/cases/common/workflows/steps/add_tags.d.ts
declare const InputSchema: z.ZodObject<{
    case_id: z.ZodString;
    tags: z.ZodArray<z.ZodString>;
}, z.core.$strip>;
declare const OutputSchema: z.ZodObject<{
    case: z.ZodObject<{
// ...
        comments: z.ZodArray<z.ZodDiscriminatedUnion<[z.ZodObject<{
// ...
            owner: z.ZodOptional<z.ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, z.core.$strip>;
}, z.core.$strip>;
type AddTagsStepInputSchema = typeof InputSchema;
type AddTagsStepOutputSchema = typeof OutputSchema;
export type AddTagsStepInput = z.infer<typeof InputSchema>;
export declare const addTagsStepCommonDefinition: CommonStepDefinition<AddTagsStepInputSchema, AddTagsStepOutputSchema>;