// x-pack/platform/plugins/shared/cases/common/workflows/steps/set_custom_field.d.ts
export declare const InputSchema: z.ZodObject<{
    case_id: z.ZodString;
    version: z.ZodOptional<z.ZodString>;
    owner: z.ZodEnum<{
        cases: "cases";
        observability: "observability";
        securitySolution: "securitySolution";
    }>;
    field_name: z.ZodString;
    value: z.ZodUnion<readonly [z.ZodString, z.ZodBoolean, z.ZodNumber, z.ZodNull]>;
}, z.core.$strip>;
export declare const OutputSchema: z.ZodObject<{
    case: z.ZodObject<{
// ...
        comments: z.ZodArray<z.ZodDiscriminatedUnion<[z.ZodObject<{
// ...
            owner: z.ZodOptional<z.ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, z.core.$strip>;
}, z.core.$strip>;
type SetCustomFieldStepInputSchema = typeof InputSchema;
type SetCustomFieldStepOutputSchema = typeof OutputSchema;
export type SetCustomFieldStepInput = z.infer<typeof InputSchema>;
export declare const setCustomFieldStepCommonDefinition: CommonStepDefinition<SetCustomFieldStepInputSchema, SetCustomFieldStepOutputSchema>;