// x-pack/platform/plugins/shared/cases/server/workflows/steps/simple_steps.d.ts
export declare const assignCaseStepDefinition: (getCasesClient: (request: KibanaRequest) => Promise<CasesClient>) => import("@kbn/workflows-extensions/server").ServerStepDefinition<import("zod").ZodObject<{
    case_id: import("zod").ZodString;
    version: import("zod").ZodOptional<import("zod").ZodString>;
    assignees: import("zod").ZodNullable<import("zod").ZodArray<import("zod").ZodObject<{
        uid: import("zod").ZodString;
    }, import("zod/v4/core").$strip>>>;
}, import("zod/v4/core").$strip>, import("zod").ZodObject<{
    case: import("zod").ZodObject<{
// ...
        comments: import("zod").ZodArray<import("zod").ZodDiscriminatedUnion<[import("zod").ZodObject<{
// ...
            owner: import("zod").ZodOptional<import("zod").ZodEnum<{
                cases: "cases";
                observability: "observability";
                securitySolution: "securitySolution";
            }>>;
// ...
        }
// ...
    }, import("zod/v4/core").$strip>;
}, import("zod/v4/core").$strip>, import("zod").ZodObject<import("zod/v4/core").$ZodLooseShape, import("zod/v4/core").$strip>>;